B-Daman Speedometer Arduino Code
Code for a speedometer for measuring the speed of a B-Daman marble:

The marble is fired along a short track. At the beginning of the track,
the marble breaks a light-beam shining on a photosensor signalling the
Arduino on pin A0 then hits a bar at the back of the track triggering a
press-switch signalling on pin A5: the speed is calculated and
displayed on a 16x1 LCD display.

By calclating the time between breaking the beam and the tripping of
the switch, and dividing that time by the length of the run, we can
determine the marble's speed.